DCON-3064: Move auditEventTransformer to dataLayer/clickHouse
Relocate from src/admin/utils/ to src/dataLayer/clickHouse/ to
co-locate with the rest of the ClickHouse AuditEvent infrastructure.
